Se diseña interfaz ejecutiva para la primera interación.

+++ src/main/java/com/autoscolombia/domain/Administrador.java
@@ -1,16 +1,15 @@
 package com.autoscolombia.domain;
 
-
-
 import jakarta.persistence.*;
 
 @Entity
 @Table(name = "administrador")
 public class Administrador {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_usuario")
-    private Long idUsuario;
+    private Integer idUsuario; // <-- antes Long
 
     @Column(name = "nombre_usuario", nullable = false, length = 50, unique = true)
     private String nombreUsuario;
@@ -18,8 +17,8 @@ public class Administrador {
     @Column(name = "contrasena_hash", nullable = false, length = 255)
     private String contrasenaHash;
 
-    public Long getIdUsuario() { return idUsuario; }
-    public void setIdUsuario(Long idUsuario) { this.idUsuario = idUsuario; }
+    public Integer getIdUsuario() { return idUsuario; }
+    public void setIdUsuario(Integer idUsuario) { this.idUsuario = idUsuario; }
 
     public String getNombreUsuario() { return nombreUsuario; }
     public void setNombreUsuario(String nombreUsuario) { this.nombreUsuario = nombreUsuario; }
