+++ src/main/java/com/autoscolombia/repository/AdministradorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AdministradorRepository extends JpaRepository<Administrador, Long> {
+public interface AdministradorRepository extends JpaRepository<Administrador, Integer> {
     Optional<Administrador> findByNombreUsuario(String nombreUsuario);
 }
